🚚 Create resource_estimation module
