--- resource_estimation/ft_optimization/mqtbench/ae_indep_qiskit_10.qasm
@@ -1,112 +1,0 @@
-// Benchmark was created by MQT Bench on 2024-03-17
-// For more information about MQT Bench, please visit https://www.cda.cit.tum.de/mqtbench/
-// MQT Bench version: 1.1.0
-// Qiskit version: 1.0.2
-
-OPENQASM 2.0;
-include "qelib1.inc";
-qreg eval[9];
-qreg q[1];
-creg meas[10];
-u2(0,-pi) eval[0];
-u2(0,-pi) eval[1];
-u2(0,-pi) eval[2];
-u2(0,-pi) eval[3];
-u2(0,-pi) eval[4];
-u2(0,-pi) eval[5];
-u2(0,-pi) eval[6];
-u2(0,-pi) eval[7];
-u2(0,-pi) eval[8];
-u3(0.9272952180016122,0,0) q[0];
-cx eval[0],q[0];
-u(-0.9272952180016122,0,0) q[0];
-cx eval[0],q[0];
-u3(0.9272952180016122,0,0) q[0];
-cx eval[1],q[0];
-u(-1.8545904360032244,0,0) q[0];
-cx eval[1],q[0];
-u3(1.8545904360032244,0,0) q[0];
-cx eval[2],q[0];
-u(-3.7091808720064487,0,0) q[0];
-cx eval[2],q[0];
-u3(2.574004435173138,-pi,-pi) q[0];
-cx eval[3],q[0];
-u(-7.4183617440128975,0,0) q[0];
-cx eval[3],q[0];
-u3(1.135176436833311,0,0) q[0];
-cx eval[4],q[0];
-u(-14.836723488025795,0,0) q[0];
-cx eval[4],q[0];
-u3(2.270352873666622,0,0) q[0];
-cx eval[5],q[0];
-u(-29.67344697605159,0,0) q[0];
-cx eval[5],q[0];
-u3(1.7424795598463425,-pi,-pi) q[0];
-cx eval[6],q[0];
-u(-59.34689395210318,0,0) q[0];
-cx eval[6],q[0];
-u3(2.7982261874869017,0,0) q[0];
-cx eval[7],q[0];
-u(-118.69378790420636,0,0) q[0];
-cx eval[7],q[0];
-u3(0.6867329322057831,-pi,-pi) q[0];
-cx eval[8],q[0];
-u(-237.38757580841272,0,0) q[0];
-cx eval[8],q[0];
-h eval[8];
-cp(-pi/2) eval[7],eval[8];
-cp(-pi/4) eval[6],eval[8];
-cp(-pi/8) eval[5],eval[8];
-cp(-pi/16) eval[4],eval[8];
-cp(-pi/32) eval[3],eval[8];
-cp(-pi/64) eval[2],eval[8];
-cp(-pi/128) eval[1],eval[8];
-cp(-pi/256) eval[0],eval[8];
-h eval[7];
-cp(-pi/2) eval[6],eval[7];
-cp(-pi/4) eval[5],eval[7];
-cp(-pi/8) eval[4],eval[7];
-cp(-pi/16) eval[3],eval[7];
-cp(-pi/32) eval[2],eval[7];
-cp(-pi/64) eval[1],eval[7];
-cp(-pi/128) eval[0],eval[7];
-h eval[6];
-cp(-pi/2) eval[5],eval[6];
-cp(-pi/4) eval[4],eval[6];
-cp(-pi/8) eval[3],eval[6];
-cp(-pi/16) eval[2],eval[6];
-cp(-pi/32) eval[1],eval[6];
-cp(-pi/64) eval[0],eval[6];
-h eval[5];
-cp(-pi/2) eval[4],eval[5];
-cp(-pi/4) eval[3],eval[5];
-cp(-pi/8) eval[2],eval[5];
-cp(-pi/16) eval[1],eval[5];
-cp(-pi/32) eval[0],eval[5];
-h eval[4];
-cp(-pi/2) eval[3],eval[4];
-cp(-pi/4) eval[2],eval[4];
-cp(-pi/8) eval[1],eval[4];
-cp(-pi/16) eval[0],eval[4];
-h eval[3];
-cp(-pi/2) eval[2],eval[3];
-cp(-pi/4) eval[1],eval[3];
-cp(-pi/8) eval[0],eval[3];
-h eval[2];
-cp(-pi/2) eval[1],eval[2];
-cp(-pi/4) eval[0],eval[2];
-h eval[1];
-cp(-pi/2) eval[0],eval[1];
-h eval[0];
-u(237.38757580841272,0,0) q[0];
-barrier eval[0],eval[1],eval[2],eval[3],eval[4],eval[5],eval[6],eval[7],eval[8],q[0];
-measure eval[0] -> meas[0];
-measure eval[1] -> meas[1];
-measure eval[2] -> meas[2];
-measure eval[3] -> meas[3];
-measure eval[4] -> meas[4];
-measure eval[5] -> meas[5];
-measure eval[6] -> meas[6];
-measure eval[7] -> meas[7];
-measure eval[8] -> meas[8];
-measure q[0] -> meas[9];
