--- resource_estimation/ft_optimization/mqtbench/grover-v-chain_indep_qiskit_7.qasm
@@ -1,328 +1,0 @@
-// Benchmark was created by MQT Bench on 2024-03-19
-// For more information about MQT Bench, please visit https://www.cda.cit.tum.de/mqtbench/
-// MQT Bench version: 1.1.0
-// Qiskit version: 1.0.2
-
-OPENQASM 2.0;
-include "qelib1.inc";
-qreg q[6];
-qreg flag[1];
-creg meas[7];
-h q[0];
-h q[1];
-h q[2];
-h q[3];
-h q[4];
-x q[5];
-cp(pi/16) q[4],q[5];
-cx q[4],q[3];
-cp(-pi/16) q[3],q[5];
-cx q[4],q[3];
-cp(pi/16) q[3],q[5];
-cx q[3],q[2];
-cp(-pi/16) q[2],q[5];
-cx q[4],q[2];
-cp(pi/16) q[2],q[5];
-cx q[3],q[2];
-cp(-pi/16) q[2],q[5];
-cx q[4],q[2];
-cp(pi/16) q[2],q[5];
-cx q[2],q[1];
-cp(-pi/16) q[1],q[5];
-cx q[4],q[1];
-cp(pi/16) q[1],q[5];
-cx q[3],q[1];
-cp(-pi/16) q[1],q[5];
-cx q[4],q[1];
-cp(pi/16) q[1],q[5];
-cx q[2],q[1];
-cp(-pi/16) q[1],q[5];
-cx q[4],q[1];
-cp(pi/16) q[1],q[5];
-cx q[3],q[1];
-cp(-pi/16) q[1],q[5];
-cx q[4],q[1];
-cp(pi/16) q[1],q[5];
-cx q[1],q[0];
-cp(-pi/16) q[0],q[5];
-cx q[4],q[0];
-cp(pi/16) q[0],q[5];
-cx q[3],q[0];
-cp(-pi/16) q[0],q[5];
-cx q[4],q[0];
-cp(pi/16) q[0],q[5];
-cx q[2],q[0];
-cp(-pi/16) q[0],q[5];
-cx q[4],q[0];
-cp(pi/16) q[0],q[5];
-cx q[3],q[0];
-cp(-pi/16) q[0],q[5];
-cx q[4],q[0];
-cp(pi/16) q[0],q[5];
-cx q[1],q[0];
-cp(-pi/16) q[0],q[5];
-u2(0,0) q[1];
-cx q[4],q[0];
-cp(pi/16) q[0],q[5];
-cx q[3],q[0];
-cp(-pi/16) q[0],q[5];
-cx q[4],q[0];
-cp(pi/16) q[0],q[5];
-cx q[2],q[0];
-cp(-pi/16) q[0],q[5];
-u2(0,0) q[2];
-cx q[4],q[0];
-cp(pi/16) q[0],q[5];
-cx q[3],q[0];
-cp(-pi/16) q[0],q[5];
-u2(0,0) q[3];
-cx q[4],q[0];
-cp(pi/16) q[0],q[5];
-u2(0,0) q[0];
-rccx q[0],q[1],q[5];
-u1(-pi) q[4];
-rccx q[2],q[5],flag[0];
-ccx q[3],flag[0],q[4];
-rccx q[2],q[5],flag[0];
-rccx q[0],q[1],q[5];
-u2(-pi,-pi) q[0];
-u2(-pi,-pi) q[1];
-u2(-pi,-pi) q[2];
-u2(-pi,-pi) q[3];
-u1(-pi) q[4];
-cp(pi/16) q[4],q[5];
-cx q[4],q[3];
-cp(-pi/16) q[3],q[5];
-cx q[4],q[3];
-cp(pi/16) q[3],q[5];
-cx q[3],q[2];
-cp(-pi/16) q[2],q[5];
-cx q[4],q[2];
-cp(pi/16) q[2],q[5];
-cx q[3],q[2];
-cp(-pi/16) q[2],q[5];
-cx q[4],q[2];
-cp(pi/16) q[2],q[5];
-cx q[2],q[1];
-cp(-pi/16) q[1],q[5];
-cx q[4],q[1];
-cp(pi/16) q[1],q[5];
-cx q[3],q[1];
-cp(-pi/16) q[1],q[5];
-cx q[4],q[1];
-cp(pi/16) q[1],q[5];
-cx q[2],q[1];
-cp(-pi/16) q[1],q[5];
-cx q[4],q[1];
-cp(pi/16) q[1],q[5];
-cx q[3],q[1];
-cp(-pi/16) q[1],q[5];
-cx q[4],q[1];
-cp(pi/16) q[1],q[5];
-cx q[1],q[0];
-cp(-pi/16) q[0],q[5];
-cx q[4],q[0];
-cp(pi/16) q[0],q[5];
-cx q[3],q[0];
-cp(-pi/16) q[0],q[5];
-cx q[4],q[0];
-cp(pi/16) q[0],q[5];
-cx q[2],q[0];
-cp(-pi/16) q[0],q[5];
-cx q[4],q[0];
-cp(pi/16) q[0],q[5];
-cx q[3],q[0];
-cp(-pi/16) q[0],q[5];
-cx q[4],q[0];
-cp(pi/16) q[0],q[5];
-cx q[1],q[0];
-cp(-pi/16) q[0],q[5];
-u2(0,0) q[1];
-cx q[4],q[0];
-cp(pi/16) q[0],q[5];
-cx q[3],q[0];
-cp(-pi/16) q[0],q[5];
-cx q[4],q[0];
-cp(pi/16) q[0],q[5];
-cx q[2],q[0];
-cp(-pi/16) q[0],q[5];
-u2(0,0) q[2];
-cx q[4],q[0];
-cp(pi/16) q[0],q[5];
-cx q[3],q[0];
-cp(-pi/16) q[0],q[5];
-u2(0,0) q[3];
-cx q[4],q[0];
-cp(pi/16) q[0],q[5];
-u2(0,0) q[0];
-rccx q[0],q[1],q[5];
-rccx q[2],q[5],flag[0];
-u1(-pi) q[4];
-ccx q[3],flag[0],q[4];
-rccx q[2],q[5],flag[0];
-rccx q[0],q[1],q[5];
-u2(-pi,-pi) q[0];
-u2(-pi,-pi) q[1];
-u2(-pi,-pi) q[2];
-u2(-pi,-pi) q[3];
-u1(-pi) q[4];
-cp(pi/16) q[4],q[5];
-cx q[4],q[3];
-cp(-pi/16) q[3],q[5];
-cx q[4],q[3];
-cp(pi/16) q[3],q[5];
-cx q[3],q[2];
-cp(-pi/16) q[2],q[5];
-cx q[4],q[2];
-cp(pi/16) q[2],q[5];
-cx q[3],q[2];
-cp(-pi/16) q[2],q[5];
-cx q[4],q[2];
-cp(pi/16) q[2],q[5];
-cx q[2],q[1];
-cp(-pi/16) q[1],q[5];
-cx q[4],q[1];
-cp(pi/16) q[1],q[5];
-cx q[3],q[1];
-cp(-pi/16) q[1],q[5];
-cx q[4],q[1];
-cp(pi/16) q[1],q[5];
-cx q[2],q[1];
-cp(-pi/16) q[1],q[5];
-cx q[4],q[1];
-cp(pi/16) q[1],q[5];
-cx q[3],q[1];
-cp(-pi/16) q[1],q[5];
-cx q[4],q[1];
-cp(pi/16) q[1],q[5];
-cx q[1],q[0];
-cp(-pi/16) q[0],q[5];
-cx q[4],q[0];
-cp(pi/16) q[0],q[5];
-cx q[3],q[0];
-cp(-pi/16) q[0],q[5];
-cx q[4],q[0];
-cp(pi/16) q[0],q[5];
-cx q[2],q[0];
-cp(-pi/16) q[0],q[5];
-cx q[4],q[0];
-cp(pi/16) q[0],q[5];
-cx q[3],q[0];
-cp(-pi/16) q[0],q[5];
-cx q[4],q[0];
-cp(pi/16) q[0],q[5];
-cx q[1],q[0];
-cp(-pi/16) q[0],q[5];
-u2(0,0) q[1];
-cx q[4],q[0];
-cp(pi/16) q[0],q[5];
-cx q[3],q[0];
-cp(-pi/16) q[0],q[5];
-cx q[4],q[0];
-cp(pi/16) q[0],q[5];
-cx q[2],q[0];
-cp(-pi/16) q[0],q[5];
-u2(0,0) q[2];
-cx q[4],q[0];
-cp(pi/16) q[0],q[5];
-cx q[3],q[0];
-cp(-pi/16) q[0],q[5];
-u2(0,0) q[3];
-cx q[4],q[0];
-cp(pi/16) q[0],q[5];
-u2(0,0) q[0];
-rccx q[0],q[1],q[5];
-rccx q[2],q[5],flag[0];
-u1(-pi) q[4];
-ccx q[3],flag[0],q[4];
-rccx q[2],q[5],flag[0];
-rccx q[0],q[1],q[5];
-u2(-pi,-pi) q[0];
-u2(-pi,-pi) q[1];
-u2(-pi,-pi) q[2];
-u2(-pi,-pi) q[3];
-u1(-pi) q[4];
-cp(pi/16) q[4],q[5];
-cx q[4],q[3];
-cp(-pi/16) q[3],q[5];
-cx q[4],q[3];
-cp(pi/16) q[3],q[5];
-cx q[3],q[2];
-cp(-pi/16) q[2],q[5];
-cx q[4],q[2];
-cp(pi/16) q[2],q[5];
-cx q[3],q[2];
-cp(-pi/16) q[2],q[5];
-cx q[4],q[2];
-cp(pi/16) q[2],q[5];
-cx q[2],q[1];
-cp(-pi/16) q[1],q[5];
-cx q[4],q[1];
-cp(pi/16) q[1],q[5];
-cx q[3],q[1];
-cp(-pi/16) q[1],q[5];
-cx q[4],q[1];
-cp(pi/16) q[1],q[5];
-cx q[2],q[1];
-cp(-pi/16) q[1],q[5];
-cx q[4],q[1];
-cp(pi/16) q[1],q[5];
-cx q[3],q[1];
-cp(-pi/16) q[1],q[5];
-cx q[4],q[1];
-cp(pi/16) q[1],q[5];
-cx q[1],q[0];
-cp(-pi/16) q[0],q[5];
-cx q[4],q[0];
-cp(pi/16) q[0],q[5];
-cx q[3],q[0];
-cp(-pi/16) q[0],q[5];
-cx q[4],q[0];
-cp(pi/16) q[0],q[5];
-cx q[2],q[0];
-cp(-pi/16) q[0],q[5];
-cx q[4],q[0];
-cp(pi/16) q[0],q[5];
-cx q[3],q[0];
-cp(-pi/16) q[0],q[5];
-cx q[4],q[0];
-cp(pi/16) q[0],q[5];
-cx q[1],q[0];
-cp(-pi/16) q[0],q[5];
-u2(0,0) q[1];
-cx q[4],q[0];
-cp(pi/16) q[0],q[5];
-cx q[3],q[0];
-cp(-pi/16) q[0],q[5];
-cx q[4],q[0];
-cp(pi/16) q[0],q[5];
-cx q[2],q[0];
-cp(-pi/16) q[0],q[5];
-u2(0,0) q[2];
-cx q[4],q[0];
-cp(pi/16) q[0],q[5];
-cx q[3],q[0];
-cp(-pi/16) q[0],q[5];
-u2(0,0) q[3];
-cx q[4],q[0];
-cp(pi/16) q[0],q[5];
-u2(0,0) q[0];
-rccx q[0],q[1],q[5];
-rccx q[2],q[5],flag[0];
-u1(-pi) q[4];
-ccx q[3],flag[0],q[4];
-rccx q[2],q[5],flag[0];
-rccx q[0],q[1],q[5];
-u2(-pi,-pi) q[0];
-u2(-pi,-pi) q[1];
-u2(-pi,-pi) q[2];
-u2(-pi,-pi) q[3];
-u1(-pi) q[4];
-barrier q[0],q[1],q[2],q[3],q[4],q[5],flag[0];
-measure q[0] -> meas[0];
-measure q[1] -> meas[1];
-measure q[2] -> meas[2];
-measure q[3] -> meas[3];
-measure q[4] -> meas[4];
-measure q[5] -> meas[5];
-measure flag[0] -> meas[6];
